% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{abbrev_colnames}
\alias{abbrev_colnames}
\title{Abbreviate column names}
\usage{
abbrev_colnames(df, maxlength)
}
\arguments{
\item{df}{dataframe}

\item{maxlength}{max length of column names after abbreviation}
}
\value{
dataframe with names os columns abbreviated
}
\description{
Abbreviate column names
}
