% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{balanced_sample}
\alias{balanced_sample}
\title{balanced sample}
\usage{
balanced_sample(df, target, n = 100)
}
\arguments{
\item{df}{dataframe}

\item{target}{name of outcome variable}

\item{n}{number of samples}
}
\value{
dataframe with balanced sample by outcome
}
\description{
Balanced Sample
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
