% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_ratio.R
\name{band_ratio}
\alias{band_ratio}
\title{band_ratio}
\usage{
band_ratio(df)
}
\arguments{
\item{df}{dataframe with bands}
}
\description{
This function calculates band ratios in the form bd = (a-b) / (a + b). The result is a dataframe containing the combinations of all the bands.
}
\details{
calculates band ratios in the form bd = (a-b) / (a + b)
}
\examples{
\dontrun{
br = band_ratio(df)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}

Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{band}
\keyword{ndvi}
\keyword{ratio}
