% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band_ratio.R
\name{band_ratio_to_raster}
\alias{band_ratio_to_raster}
\title{band_ratio_to_Raster}
\usage{
band_ratio_to_raster(raster_ext, raster_input_path, raster_output_path)
}
\arguments{
\item{raster_ext}{extension of input raster file (.asc, .tif)}

\item{raster_input_path}{path of input raster files}

\item{raster_output_path}{path of output raster files}
}
\description{
This function calculates band ratios in the form bd = (a-b) / (a + b). The result is write in a raster file with the same extension of input files
}
\details{
calculates band ratios in the form bd = (a-b) / (a + b)
}
\examples{
\dontrun{
raster_input_path = "./asc"
raster_ext =  ".asc"
raster_output_path = "./band_ratio"
band_ratio_to_raster(raster_ext,raster_input_path,raster_output_path )
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}

Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{band}
\keyword{ndvi}
\keyword{ratio}
