% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_stock.R
\docType{data}
\name{c_stock}
\alias{c_stock}
\title{Brazilian Soil Stock Carbon Data}
\format{A dataframe with 8156 rows and 67 columns}
\usage{
data(c_stock)
}
\description{
This data comes Brazilian data soil carbon stock
}
\examples{
\dontrun{
df = c_stock
}
}
\keyword{datasets}
