% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean columns names}
\usage{
clean_names(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function cleans dataframe columns names removing especial symbols and blank space
}
\details{
details
}
\examples{
\dontrun{
clean_names(df)
}
}
\keyword{Clean}
\keyword{columns}
\keyword{names}
