% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{comma_to_point}
\alias{comma_to_point}
\title{Convert comma to point}
\usage{
comma_to_point(x, tonumeric = TRUE)
}
\arguments{
\item{x}{vector}

\item{tonumeric}{logical TRUE transform to numeric after change comma to point}
}
\description{
Convert comma to point
}
