% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_validation_poligon.R
\name{create_folds_poligon}
\alias{create_folds_poligon}
\title{This function  creates folds  for samples collected in the form of polygons.In this case
all samples inside a polygon are placed in a fold avoiding the division in multiples
folds..}
\usage{
create_folds_poligon(outcome, polig, nfolds = 10, seeds = 313)
}
\arguments{
\item{outcome}{vector with  outcome variable}

\item{nfolds}{number of folds to be created}

\item{poligon}{vector with  poligon id}

\item{seed}{numeric seed to control random numbers generation}
}
\description{
This function  creates folds  for samples collected in the form of polygons.In this case
all samples inside a polygon are placed in a fold avoiding the division in multiples
folds..
}
\examples{
\dontrun{
dfall = df \%>\% select(classe, poligono, everything())
index = create_folds_poligon(dfall, class = classe, poligon = poligono, nfolds = 5, seed = 123)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
