% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_preparation}
\alias{data_preparation}
\title{This function preprare data to analisys using vtreat library}
\usage{
data_preparation(df, prune = 0.99, p = 0.75, seed = 313)
}
\arguments{
\item{df}{global dataframe}

\item{prune}{suppress variables with significance above this level}

\item{p}{the percentage of data that goes to training}

\item{seed}{seed}
}
\description{
This function preprare data to analisys using vtreat library
}
\examples{
\dontrun{
dp = data_preparation(df = c_stock, p = 0.75, prune = 0.99)
treino = dp$treino
teste = dp$teste
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{data}
\keyword{prepartion}
\keyword{vtreat}
