% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_categoric.R
\name{data_preparation_categoric}
\alias{data_preparation_categoric}
\title{This function preprare data categoric to analisys using vtreat library}
\usage{
data_preparation_categoric(df, p = 0.75, positivo = 1, prune = 0.99,
  seed = 313)
}
\arguments{
\item{df}{global dataframe}

\item{p}{the percentage of data that goes to training}

\item{positivo}{Value/level of outcome to be considered "success"}

\item{prune}{suppress variables with significance above this level}

\item{seed}{seed}
}
\description{
This function preprare data categoric to analisys using vtreat library
}
\details{
details
}
\examples{
\dontrun{
data_prepatarion_categoric(df, 0.75)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{data}
\keyword{prepartion}
\keyword{vtreat}
