% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_to_raster.R
\name{dataframe_to_raster}
\alias{dataframe_to_raster}
\title{dataframe_to_raster}
\usage{
dataframe_to_raster(df)
}
\arguments{
\item{df}{dataframe in xYZ format where x and y are coordinates and z is the variable to create raster.}
}
\description{
This function convertes a dataframe in XYZ format, to a raster
}
\details{
details
}
\examples{
\dontrun{
dataframe_to_raster(df)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{correlation}
\keyword{dataframe}
\keyword{raster}
