% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stat_polygon.R
\name{descriptive_stat_polygon}
\alias{descriptive_stat_polygon}
\title{Descriptive statistics polygon}
\usage{
descriptive_stat_polygon(df, poligon_id)
}
\arguments{
\item{df}{dataframe}

\item{poligon_id}{name of column thats contains polygon id}
}
\description{
This function calculates Descriptive statistics for samples collected in polygon
}
\details{
calculates Descriptive statistics for samples collected in polygon
}
\examples{
\dontrun{
b = descriptive_stat_polygon(df,"id")
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{polygon}
\keyword{statics}
