% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_predict_to_raster.R
\name{df_predict_to_raster}
\alias{df_predict_to_raster}
\title{This function creates a raster as a result of predicting an adjusted model, given x and y, and a dataframe containing the data needed to predict. The result is a raster map.}
\usage{
df_predict_to_raster(fit, xy, df)
}
\arguments{
\item{fit}{model}

\item{xy}{dataframe with x and y coordinates. Names of column must be "x" and "y"}

\item{df}{dataframe with data neeeded to predict the model. Names of columns must be the same of vars that was be used to train the model}
}
\description{
This function creates a raster as a result of predicting an adjusted model, given x and y, and a dataframe containing the data needed to predict. The result is a raster map.
}
\details{
details
}
\examples{
\dontrun{
ptr = df_predict_to_raster(fit, xy, df)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{map,}
\keyword{predict,}
\keyword{raster}
