% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_fit.R
\name{easy_fit}
\alias{easy_fit}
\title{This function performs the training of the chosen models to a group of outcomes (y)}
\usage{
easy_fit(dy, dx, nfolds = 10, repeats = 1, metric = ifelse(is.factor(dy[,
  1]), "Kappa", "Rsquared"), cpu_cores = 7, tune_length = 5,
  fun_RFE = rfFuncs, repeats_RFE = 1, metric_RFE = ifelse(is.factor(dy[,
  1]), "Kappa", "Rsquared"), nfolds_RFE = 5, sizes_RFE = c(2:10, 15),
  tolerance_RFE = 0, models = c("ridge", "rf", "cubist", "pls", "foba",
  "gbm", "glmboost"), preprocess = NULL, seeds = NULL)
}
\arguments{
\item{dy}{outcomes}

\item{dx}{co-variables}

\item{nfolds}{Number of folds to be build in crossvalidation}

\item{repeats}{repeats}

\item{metric}{metric used to evaluate model fit. For numeric outcome ("RMSE", "Rsquared)}

\item{cpu_cores}{Number of CPU cores to be used in parallel processing}

\item{tune_length}{By default, this argument is the number of levels for each tuning parameters that should be generated by train.}

\item{fun_RFE}{A list of functions for model fitting, prediction and variable importance}

\item{repeats_RFE}{repeats to recursive feature elimination}

\item{metric_RFE}{metric to recursive feature elimination}

\item{nfolds_RFE}{nfolds to recursive feature elimination}

\item{sizes_RFE}{A numeric vector of integers corresponding to the number of features that should be retained}

\item{tolerance_RFE}{tolerance}

\item{models}{vector of models to be fitted. Uses  algortims names from Caret package.}

\item{preprocess}{pre process}

\item{seeds}{seeds}
}
\description{
This function performs the training of the chosen models to a group of outcomes (y)
}
\details{
details
}
\examples{
\dontrun{
easy_fit(dy,dx,train,models = c("ridge", "rf", "cubist"), metric = "Rsquared", tolerance = 2)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{Train}
\keyword{kappa}
