% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{eliminate_symbol}
\alias{eliminate_symbol}
\title{eliminate symbols from string}
\usage{
eliminate_symbol(vx, replace = " ")
}
\arguments{
\item{vx}{character vector}

\item{replace}{symbol the will replace eliminated symbol character}
}
\description{
Eliminate symbol
}
