% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_point.R
\name{extract_data_point}
\alias{extract_data_point}
\title{extract_data_point}
\usage{
extract_data_point(path_raster, raster_type = "*.asc", path_point,
  name_point, cpu_cores = 1, remove_NA = TRUE)
}
\arguments{
\item{path_raster}{Path para a pasta onde estão os rasters}

\item{raster_type}{Tipo do arquivo raster a ser lido (*.tif, *.asc, *.img ... )}

\item{path_point}{Path para a pasta onde esta os pontos de amostragem}

\item{name_point}{name of shapefile of points without extension}

\item{cpu_cores}{number of cpu cores used in processing}

\item{remove_NA}{remove NA values after extraction}
}
\description{
This function allows extract data from raster layers from point
}
\details{
details
}
\examples{
\dontrun{
extract_data_point(pathraster,"*.tif", path_point, "amostras_cafe")
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{data}
\keyword{extract}
\keyword{point}
\keyword{raster}
