% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_detect.R
\name{factor_detect}
\alias{factor_detect}
\title{factor_detect}
\usage{
factor_detect(df, max.unique = 20)
}
\arguments{
\item{df}{dataframe}

\item{max.unique}{number maximum of unique values to be considered as a factor varaiable}
}
\description{
This function detect factor variables
}
\details{
details
}
\examples{
\dontrun{
factor_detect (df, max.unique = 10)
}
}
\keyword{detect}
\keyword{factor}
