% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_poligon.R
\name{fold_poligon}
\alias{fold_poligon}
\title{fold poligon - separete samples of whole polygons to cross validation}
\usage{
fold_poligon(nfold, poligon, seed = 123)
}
\arguments{
\item{nfold}{Number of folds to be created}

\item{poligon}{Vector with the sampling polygons}

\item{seed}{number tp seed generator}
}
\value{
A list with folds;. Each fold contains line number of data to be uses in cross validation
}
\description{
This function separates samples into polygons for cross-validation so that all pixels of a
polygon are allocated to a single fold. This avoids overfitting of the result due to spatial
correlation present in geographically close samples.
}
\details{
details
}
\examples{
\dontrun{
fold_poligon(nfold = 10, poligon =df$poligon)
}
}
