% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getdist_rectangle}
\alias{getdist_rectangle}
\title{calculates distance from points}
\usage{
getdist_rectangle(px, py, nx, ny)
}
\arguments{
\item{px}{vector of x coordinates}

\item{py}{vector of y coordinates}

\item{nx}{number of divisions in x axis}

\item{ny}{number of divisions in y axis}
}
\description{
Calculates distance from points
}
