% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{grid_dist_rasterize}
\alias{grid_dist_rasterize}
\title{Grid Distance rasterize}
\usage{
grid_dist_rasterize(fileraster, path_output = ".", nx = 3, ny = 3,
  format = ".asc")
}
\arguments{
\item{fileraster}{raster file raster to extract bounds, resolution and projection information}

\item{path_output}{path to output rasters files}

\item{nx}{number of divisions in x axis}

\item{ny}{number of divisions in y axis}

\item{format}{format of raster files ('.asc', '.tif', '.grd', '.sdat'..etc). See Help of writeRaster function of raster package to more formats}
}
\description{
Create rasters of distances to a grid
}
\details{
grid distance rasterize
}
\examples{
\dontrun{0
grid_rasterize('bio_14.asc', path_output = '.', nx = 3, ny = 4, format = '.asc')
}
}
