% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halarick_descriptor.R
\name{halarick_descriptor}
\alias{halarick_descriptor}
\title{halarick_descriptor}
\usage{
halarick_descriptor(band, ngrey = 16, window = c(3, 3), shift = list(c(0,
  1), c(1, 1), c(1, 0), c(1, -1)), statistics = c("mean", "mean_ENVI",
  "variance", "variance_ENVI", "homogeneity", "contrast", "dissimilarity",
  "entropy", "second_moment", "correlation"))
}
\arguments{
\item{band}{raster with a single band}

\item{ngrey}{number of grey levels to use in texture calculation}

\item{window}{2 element list with row and column dimensions of the texture window}

\item{shift}{a matrix where each row gives an (x, y) shift to use when computing co-occurrency matrices. Textures will be calculated for each shift, and the average over all shifts will be returned.}

\item{statistics}{a list of strings naming the texture statistics to calculate}
}
\description{
This function calculates halarick texture descriptors
}
\details{
details
}
\examples{
\dontrun{
descriptor = halarick_descriptor(band4)
plot(descriptor)
}
}
\keyword{halaric}
