% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_correlation.R
\name{high_correlation}
\alias{high_correlation}
\title{high_correlation}
\usage{
high_correlation(df, cutoff = 0.95, correl = "linear", verbose = FALSE)
}
\arguments{
\item{df}{input dataframe}

\item{cutoff}{numeric value for the pair-wise absolute correlation cutoff}

\item{correl}{type of correlation. linear ou rdc}

\item{verbose}{A boolean for printing the details and plot graphic}
}
\description{
This function searches through a correlation matrix and returns a vector of integers corresponding to columns to remove to reduce pair-wise correlations.
}
\details{
details
}
\examples{
\dontrun{
high_correlation(df, 0.95, correl = "linear", TRUE)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{correlated}
\keyword{high}
