% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriga.R
\name{kriga}
\alias{kriga}
\title{kriga - Automatric krige}
\usage{
kriga(df, target_var, nrep = 10, px, py, p = 0.75, seed = NULL)
}
\arguments{
\item{df}{dataframe with targer variable, coordinate x  and coordinate y}

\item{target_var}{Name of column with target_var}

\item{nrep}{Number of repetitions}

\item{px}{names of column with coordinate x}

\item{py}{names of column with coordinate y}

\item{p}{proportion between training and validation}

\item{seed}{seed number to allow reprodutibility}
}
\description{
kriga - Automatric krige
}
\details{
details
}
\examples{
\dontrun{
result.krig = kriga(df = d,target_var = "argila", px = 'point_x',
                    py = 'point_y',nrep = 100, p = 0.75 , seed = 313)
media = result.krig \%>\% select(-(model:repet))
        \%>\% summarise_all(funs(med = mean, desvpad = sd ))
print(knitr::kable(media))
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{geostatics}
\keyword{krige}
