% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse.R
\docType{data}
\name{landuse}
\alias{landuse}
\title{Land use data}
\format{A dataframe with 465 rows and 99 columns}
\usage{
data(landuse)
}
\description{
Land use class, terrain and satellite image covariates
}
\examples{
\dontrun{
df = landuse
}
}
\keyword{datasets}
