% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda_median.R
\name{moda_median}
\alias{moda_median}
\title{moda_median}
\usage{
moda_median(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function calculates moda or median for each column of a dataframe
}
\details{
details
}
\examples{
\dontrun{
moda_median(df)
}
}
\keyword{median}
\keyword{moda}
