% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_performance.R
\name{model_performance}
\alias{model_performance}
\title{Model_Performance}
\usage{
model_performance(list_fit)
}
\arguments{
\item{list_fit}{list generated by funciot easy_model}
}
\description{
This function performs the training of the chosen classifier
}
\details{
details
}
\examples{
\dontrun{
dfperf = model_perfomance(l.fit)
}
}
