% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{model_tags}
\alias{model_tags}
\title{models tags}
\usage{
model_tags(is.regression = TRUE)
}
\arguments{
\item{is.regression}{if TRUE return regression models tags if FALSE return classification models tags}
}
\description{
model_tags : return list of models tags supported by caret package
}
