% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{moments - This function calculates the univariate statistics of a variable in vector form}
\usage{
moments(x, plot = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{vetor wiht data}

\item{plot}{define if plot the graphics}

\item{na.rm}{remove nodata before calcululate moments}
}
\description{
code from moments package https://cran.r-project.org/web/packages/moments/index.html
}
\details{
details
}
\examples{
\dontrun{
vt = c(12,22,34,26,5,19)
print(moments(v))
}
}
