% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nasa.r
\docType{data}
\name{nasa}
\alias{nasa}
\title{NASA spatio-temporal data}
\format{A [tbl_cube] with 41,472 observations.}
\usage{
nasa
}
\description{
This data comes from the ASA 2007 data expo,
\url{http://stat-computing.org/dataexpo/2006/}. The data are geographic and
atmospheric measures on a very coarse 24 by 24 grid covering Central
America. The variables are: temperature (surface and air), ozone,
air pressure, and cloud cover (low, mid, and high). All variables are
monthly averages, with observations for Jan 1995 to Dec 2000. These data
were obtained from the NASA Langley Research Center Atmospheric Sciences
Data Center (with permission; see important copyright terms below).
}
\section{Dimensions}{


\itemize{
  \item `lat`, `long`: latitude and longitude
  \item `year`, `month`: month and year
}
}

\section{Measures}{


\itemize{
  \item `cloudlow`, `cloudmed`, `cloudhigh`: cloud cover
    at three heights
  \item `ozone`
  \item `surftemp` and `temperature`
  \item `pressure`
}
}

\examples{
nasa
}
