% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_correl.R
\name{non_linear_correl}
\alias{non_linear_correl}
\title{non_linear_correl}
\usage{
non_linear_correl(df)
}
\arguments{
\item{df}{dataframe with numerical data}
}
\description{
This function calculates non linear correlation between variables using rdc algorithm
}
\details{
details
}
\examples{
\dontrun{
non_linear_correl(df)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{RDC}
\keyword{correlation}
\keyword{linera}
\keyword{non}
