% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_performance.R
\name{null_model}
\alias{null_model}
\title{fit null model}
\usage{
null_model(train_y, test_y)
}
\arguments{
\item{train_y}{vector with train outcome variable}

\item{test_y}{vector with test outcome variable}
}
\description{
calculate RMSE of null model
}
\examples{
\dontrun{
null_model(train_y = train$Ozone, test_y = test$Ozone)
}
}
