% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{numeric_to_factor}
\alias{numeric_to_factor}
\title{Numeric to Factor}
\usage{
numeric_to_factor(vx, npercentil = 4)
}
\arguments{
\item{vx}{vector of numeric}

\item{percentil}{numeric the number of percintil to be splited the vector vx}
}
\value{
vector of factors
}
\description{
convert numeric data to factor
}
