% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_performance.R
\name{plot_confusio_matrix}
\alias{plot_confusio_matrix}
\title{Plot Confusion Matrix}
\usage{
plot_confusio_matrix(obs, pred)
}
\arguments{
\item{obs}{observed values}

\item{pred}{predicted values}
}
\description{
This function  Plot Confusion Matrix
}
\details{
Plot Confusion Matrix
}
