% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_performance.R
\name{plot_predict_observed_residual}
\alias{plot_predict_observed_residual}
\title{Plot predict observed and residual}
\usage{
plot_predict_observed_residual(result, residual = FALSE)
}
\arguments{
\item{result}{return of function rum_model_performance}

\item{residual}{logic value indicates if plot or not residuals}
}
\description{
This function plots predict,  observed values and residual
}
\details{
Plot predict observed and residual

details
}
\examples{
\dontrun{
plot_predict_observed_residual(result, TRUE)
}
}
