% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{point_to_camel}
\alias{point_to_camel}
\title{Convert points to uppercase letter in a string}
\usage{
point_to_camel(vx)
}
\arguments{
\item{vx}{character vector}
}
\value{
vector with points replaced by uppercase character
}
\description{
Convert points to uppercase letter in a string
}
