% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_big_raster.R
\name{predict_big_raster}
\alias{predict_big_raster}
\title{Title Predict big raster}
\usage{
predict_big_raster(model, st, path_file, suffix = "etp_2000_",
  format = "GTiff", tiles = 5, cpu_cores = 7)
}
\arguments{
\item{model}{model to be predicted}

\item{st}{stack with co-variable}

\item{path_file}{path to storage results}

\item{suffix}{suffix to be adictioned to name of file results}

\item{format}{graphics format of result file}

\item{tiles}{number of  divisions in x and y to create tiles}

\item{cpu_cores}{numeber of cpu cores}
}
\description{
Title Predict big raster
}
\examples{
predict_big_raster(model = modelo_calibra, st = stcovar,
                   path_file ='./mosaic/', format = "GTiff",  s
                   uffix = 'etp_', tiles = 3, cpu_cores = 7 )
}
