% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_stack_to_map_parallel}
\alias{predict_stack_to_map_parallel}
\title{predict stack to map parallel}
\usage{
predict_stack_to_map_parallel(model_list, cpu_cores = 2, stack_var,
  path_result, namefile, result_type = ".tif")
}
\arguments{
\item{model_list}{list of fit models by function run_models}

\item{stack_var}{Stack with all vars used to fit them model}

\item{path_result}{path to store maps with results of prediction}

\item{namefile}{sufix of name of file with results of prediction}

\item{result_type}{type of file (extension) to be saved after prediction}

\item{cou_cores}{number of cpu_cores}
}
\description{
This function predict a model fit and save result as a map raster file in paralllel processing
}
\details{
details
}
\examples{
\dontrun{
predict_stack_to_map_paralllel(fit.models,
              cpu_cores = 2,
              stack_var = predicao,
              path_result = './result/',
              namefile = 'uso_',
              result_type = ".tif")
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{maps}
\keyword{models}
\keyword{predict}
\keyword{raster}
