% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_to_map}
\alias{predict_to_map}
\title{predict to map}
\usage{
predict_to_map(model_list, path_raster, raster_type = ".asc",
  path_result, namefile, result_type = ".tif")
}
\arguments{
\item{model_list}{list of fit models by function run_models}

\item{path_result}{path to store maps with results of prediction}

\item{namefile}{sufix of name of file with results of prediction}

\item{result_type}{type of file (extension) to be saved after prediction}
}
\description{
This function predict a model fit and save result as a map raster file
}
\details{
details
}
\examples{
\dontrun{
predict_to_map(fit.models,
              path_raster = './covar_asc/',
              raster_type = ".asc",
              path_result = './result/',
              namefile = 'uso_',
              result_type = ".tif")
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{maps}
\keyword{models}
\keyword{predict}
\keyword{raster}
