% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rasterize_coordinates}
\alias{rasterize_coordinates}
\title{Rasterize coordinates}
\usage{
rasterize_coordinates(fileraster, path_output = ".",
  prefix_name = "coord_", format = ".asc")
}
\arguments{
\item{fileraster}{raster file raster to extract bounds, resolution and projection information}

\item{path_output}{path to output rasters files}

\item{format}{format of raster files ('.asc', '.tif', '.grd', '.sdat'..etc). See Help of writeRaster function of raster package to more formats}

\item{prefix_nama}{prefix to use to create name of rasters files}
}
\description{
Create rasters of coordinates x and y
}
\details{
Rasterize coordinates
}
\examples{
\dontrun{0
rasterize_coordinates(fileraster = './raster/bio_14.asc', path_output = './raster', prefix_name = 'coord_', format = '.tif')
}
}
