% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_feature_elimination.R
\name{recursive_feature_elimination}
\alias{recursive_feature_elimination}
\title{This function performs an recursive feature elimination using the rfe function from caret package}
\usage{
recursive_feature_elimination(df, sizes = c(2:5, 10), index = NULL,
  nfolds = 5, fun = rfFuncs, cpu_cores = 6,
  metric = ifelse(is.factor(df[, 1]), "Kappa", "Rsquared"), seeds = NULL,
  verbose = TRUE)
}
\arguments{
\item{df}{dataframe, with income variable in the first column}

\item{sizes}{A numeric vector of integers corresponding to the number of features that should be retained}

\item{index}{Users cross validation folds. Default = NULL}

\item{nfolds}{Number of folds to be build in cross-validation}

\item{fun}{A list of functions for model fitting, prediction and variable importance}

\item{cpu_cores}{Number of CPU cores to be used in parallel processing}

\item{metric}{metric used to evaluate model fit. For numeric outcome ("RMSE", "Rsquared) and for categorical outcome ("Accuracy","Kappa")}

\item{seeds}{seeds}

\item{verbose}{verbose}
}
\description{
This function performs an recursive feature elimination using the rfe function from caret package
}
\details{
details
}
\examples{
\dontrun{
rfe(df,sizes = c(2:10,12,15,20), cpu_cores = 4)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{Elimination}
\keyword{Feature}
\keyword{Recursive}
