% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{This function performs the training of the chosen regressor}
\usage{
regression(df.train, formula = NULL, preprocess = NULL, regressor = "rf",
  resample_ = "cv", nfolds = 10, repeats = NA, index = NULL,
  cpu_cores = 0, tune_length = 5, metric = "Rsquared", seeds = NULL,
  verbose = FALSE)
}
\arguments{
\item{df.train}{Training dataframe}

\item{formula}{A formula of the form y ~ x1 + x2 + ... If users don't inform formula, the first column will be used as Y values and the others columns with x1,x2....xn}

\item{preprocess}{pre process}

\item{regressor}{Choice of regressor to be used to train model. Uses  algortims names from Caret package.}

\item{resample_}{ressample method 'boot', 'boot632', 'optimism_boot', 'boot_all', 'cv', 'repeatedcv', 'LOOCV', 'LGOCV','none', 'oob', 'timeslice', 'adaptive_cv', 'adaptive_boot', 'adaptive_LGOCV'}

\item{nfolds}{Number of folds to be build in crossvalidation}

\item{repeats}{repeats}

\item{index}{index}

\item{cpu_cores}{Number of CPU cores to be used in parallel processing}

\item{tune_length}{This argument is the number of levels for each tuning parameters that should be generated by train}

\item{metric}{metric used to evaluate model fit. For numeric outcome ("RMSE", "Rsquared)}

\item{seeds}{seeds}

\item{verbose}{verbose}
}
\description{
This function performs the training of the chosen regressor
}
\details{
details
}
\examples{
\dontrun{
regression(df.train = df, regressor = "rf", metric = "Rsquared", seeds = 313)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{RMSE}
\keyword{Rsquared}
\keyword{Train}
\keyword{regression}
