% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{remove_accentuation}
\alias{remove_accentuation}
\title{Remove acentuation}
\usage{
remove_accentuation(vx)
}
\arguments{
\item{vx}{character vector}
}
\value{
vector of character whitout acentuation
}
\description{
Remove acentuation
}
