% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{remove_first_symbol}
\alias{remove_first_symbol}
\title{Remove first symbol}
\usage{
remove_first_symbol(vx)
}
\arguments{
\item{vx}{vector of character}
}
\value{
vector of character with first symbol removed
}
\description{
Remove symbol when it is the first character of a string
}
