% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{remove_var_na}
\alias{remove_var_na}
\title{Remove Columns with too many NA values}
\usage{
remove_var_na(df, d_count_na, tolerance = 75, verbose = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{d_count_na}{dataframe with NA information created by na.Count}

\item{tolerance}{percentual value of tolerance of NA values in a variable}

\item{verbose}{verbose}
}
\description{
This function Remove Columns with too many NA values
}
\details{
details
}
\examples{
\dontrun{
remove_var_na(df, d.count.na, tolerance = 30, verbose = TRUE)
}
}
