% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_feature_elimination.R
\name{rfe_result}
\alias{rfe_result}
\title{Recursive Feature Elimination plot results}
\usage{
rfe_result(fit.rfe)
}
\arguments{
\item{fit.rfe}{Results from recursive_feature_elimination  function}
}
\description{
This function plot results of Recursive Feature Elimination (RFE)
}
\details{
details
}
\examples{
\dontrun{
rfe_result(fit.rfe)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{Elimination}
\keyword{Feature}
\keyword{Recursive}
\keyword{plot}
\keyword{results}
