% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{roi}
\alias{roi}
\title{Create polygon of Region of Interest (ROI)
code by Jeffrey Evans}
\usage{
roi(extent, project)
}
\arguments{
\item{extent}{projection of region of  interest}

\item{project}{projection string (ex: "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")}
}
\description{
Create polygon of Region of Interest (ROI)
code by Jeffrey Evans
}
\examples{
\dontrun{
ext <- as(raster::extent(78.46801, 78.83157, 19.53407, 19.74557), "SpatialPolygons")
proj <- "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs"
pol.roi = roio(ext, proj)
}
}
