% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models.R
\name{run_models}
\alias{run_models}
\title{Run Models}
\usage{
run_models(df, models = ifelse(is.factor(df[, 1]), c("qda", "rf", "gbm",
  "C5.0"), c("lm", "cubist", "gbm", "rf")), formula = NULL,
  preprocess = NULL, index = NULL, rsample = "cv", nfolds = 10,
  repeats = NA, tune_length = 5, cpu_cores = 0,
  metric = ifelse(is.factor(df[, 1]), "Kappa", "Rsquared"), seeds = NULL,
  verbose = TRUE)
}
\arguments{
\item{df}{Training dataframe}

\item{models}{chosen models to be used to train model. Uses  algortims names from Caret package.}

\item{formula}{A formula of the form y ~ x1 + x2 + ... If users don't inform formula, the first column will be used as Y values and the others columns with x1,x2....xn}

\item{preprocess}{pre process}

\item{index}{Users cross validation folds. Default = NULL}

\item{rsample}{resample method 'boot', 'boot632', 'optimism_boot', 'boot_all', 'cv', 'repeatedcv', 'LOOCV', 'LGOCV','none', 'oob', 'timeslice', 'adaptive_cv', 'adaptive_boot', 'adaptive_LGOCV'}

\item{nfolds}{Number of folds to be build in crossvalidation}

\item{repeats}{number of repeats to resample method repeatedcv}

\item{tune_length}{This argument is the number of levels for each tuning parameters that should be generated by train}

\item{cpu_cores}{Number of CPU cores to be used in parallel processing}

\item{metric}{metric used to evaluate model fit. For numeric outcome ('RMSE', 'Rsquared)}

\item{seeds}{generate random seeds to allow reproductible results}

\item{verbose}{prints results during the execution of the function}
}
\description{
This function run many models using the same data
}
\details{
details
}
\examples{
\dontrun{
models = c('ridge', 'rf', 'cubist','pls','pcr','foba','gbm','glmboost')
fit_models = run_models(df,models = models)
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{models}
