% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{save_gggraphics}
\alias{save_gggraphics}
\title{Save ggplot graphics}
\usage{
save_gggraphics(object, graphic_format = c("jpg", "png"), file_path,
  width = 15, height = 15, units = "cm", dpi = 100)
}
\arguments{
\item{object}{ggplot graphic}

\item{graphic_format}{format to be saved ("jpg", "png")}

\item{file_path}{path and file name of graphic arquive to be saved}

\item{width}{width of graphic in units of measure}

\item{height}{height  of graphic in units of measure}

\item{units}{unit of measure ("cm", "inch")}

\item{dpi}{dot per inch resoluition}
}
