% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{select_model}
\alias{select_model}
\title{select model by tag}
\usage{
select_model(is.regression = TRUE, vtags = c("Boosting", "Bagging"))
}
\arguments{
\item{is.regression}{if TRUE return regression models tags if FALSE return classification models tags}

\item{vtags}{vetctor of taga used in model selection}
}
