% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{separability.measures}
\alias{separability.measures}
\title{separability.measures}
\usage{
separability.measures(Vector.1, Vector.2)
}
\arguments{
\item{Vector.1}{data of class 1}

\item{Vector.2}{data of class 2}
}
\description{
This function calculates Jeffries-Matusita distance, Bhattacharya distance and the transformed divergence.
}
\details{
details
}
\examples{
\dontrun{
sample.1 <- c (1362, 1411, 1457, 1735, 1621, 1621, 1791, 1863, 1863, 1838)
sample.2 <- c (1362, 1411, 1457, 10030, 1621, 1621, 1791, 1863, 1863, 1838)

#separability between these two samples
separability.measures ( sample.1 , sample.2 )
}
}
\references{
https://stats.stackexchange.com/questions/78849/measure-for-separability
}
\keyword{distance}
\keyword{divergence}
