% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{space_to_symbol}
\alias{space_to_symbol}
\title{space to symbol}
\usage{
space_to_symbol(vx, symbol)
}
\arguments{
\item{vx}{character vector}

\item{symbol}{symbol the will replace space character}
}
\description{
Convert string space to any symbol
}
