% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{spaceless}
\alias{spaceless}
\title{spaceless}
\usage{
spaceless(vx, symbol = "-")
}
\arguments{
\item{vx}{character vector}

\item{symbol}{symbol or character to replace space}
}
\description{
remove spaces from a character vector
}
