% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_boxplot.R
\name{stack_boxplot}
\alias{stack_boxplot}
\title{stack_boxplot}
\usage{
stack_boxplot(x, y, y.limit, name1, name2, color1, color2, title)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{y.limit}{y.limit}

\item{name1}{name1}

\item{name2}{name2}

\item{color1}{color1}

\item{color2}{color2}

\item{title}{title}
}
\description{
This function plot two variables
}
\details{
details
}
\examples{
\dontrun{
stack_boxplot(b1, b2, 2000, "banda1", "banda2","red","blue", "band1 x band2")
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{boxplot}
