% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_SNE.R
\name{t_SNE_plot}
\alias{t_SNE_plot}
\title{t_SNE_plot dimension reduction}
\usage{
t_SNE_plot(dx, y)
}
\arguments{
\item{dx}{covariates}

\item{y}{responde variable}
}
\description{
t-SNE is a method for constructing a low dimensional embedding of high-dimensional data, distances or similarities.
}
\details{
details
}
\examples{
\dontrun{
dx = dxy[,-1]
dy = dxy[,1]
t_SNE_plot(dx,dy)
}
}
\keyword{dimension}
\keyword{reduction}
\keyword{t_SNE}
