% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{train_test}
\alias{train_test}
\title{train test split}
\usage{
train_test(df, p = 0.75, groups = 10, seed = NULL)
}
\arguments{
\item{df}{dataframe}

\item{p}{proportion between  train and test sets}

\item{groups}{numbre of groups to be used in stratified sample}
}
