% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{train_validation_perfil}
\alias{train_validation_perfil}
\title{Separete Train and test samples whit group variable}
\usage{
train_validation_perfil(df, perfil, p = 0.8, seed = NULL)
}
\arguments{
\item{df}{dataframe}

\item{p}{sample proportion}

\item{seed}{random number seed}

\item{perfilt}{groupp vector}
}
\description{
Separete Train and test samples whit group variable
}
\examples{
\dontrun{
vs = train_validatio_perfil(d, d$perfil, p = 0.8, seed = 313)
}
}
