% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_validation_poligon.R
\name{train_validation_poligon}
\alias{train_validation_poligon}
\title{This function performs the separation of training samples and validation for samples collected in the form of polygons.
Note that the class has to be the first column and the polygon number has to be the second column of the dataframe.}
\usage{
train_validation_poligon(dfall, p = 0.75, seed = 313)
}
\arguments{
\item{dfall}{Dataframe with class and poligon columns}

\item{p}{the percentage of data that goes to training}

\item{seed}{seed to control random numbers generation}
}
\description{
This function performs the separation of training samples and validation for samples collected in the form of polygons.
Note that the class has to be the first column and the polygon number has to be the second column of the dataframe.
}
\examples{
\dontrun{
dfall = df \%>\% select(classe, poligono, everything())
vt = treino_validacao_poligono(dfall, p = 0.80, seed = 123)
treino = dfall[treino,]
valida = dfall[-treino,]
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
