% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_count.r
\name{unique_count}
\alias{unique_count}
\title{unique_count}
\usage{
unique_count(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function counts unique values by  column and calculate relative value of uniquiness
}
\details{
details
}
\examples{
\dontrun{
unique_count(df)
}
}
\keyword{counts}
\keyword{uniquevalues}
