% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stat_polygon.R
\name{descriptive_stat_polygon}
\alias{descriptive_stat_polygon}
\title{Descriptive statistics polygon}
\usage{
descriptive_stat_polygon(df, poligon_id, desc = c("min", "mean", "max", "sd"))
}
\arguments{
\item{df}{dataframe}

\item{poligon_id}{name of column thats contains polygon id}

\item{desc}{parameters os descritive statistics to be calculated.Choices are : c("min", "median", "mean", "max", "sd","skewness", "kurtosis", "p1", "p10", "p25","p33", "p66", "p75", "p90", "p99")}
}
\description{
This function calculates Descriptive statistics for samples collected in polygon
}
\details{
calculates Descriptive statistics for samples collected in polygon
}
\examples{
\dontrun{
b = descriptive_stat_polygon(df,"id")
}
}
\author{
Elpidio Filho, \email{elpidio@ufv.br}
}
\keyword{polygon}
\keyword{statics}
