% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{na_count}
\usage{
na_count(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function counts NA by column and calculate relative value of NA
}
\details{
details
}
\examples{
\dontrun{
na_count(df)
}
}
\keyword{NA}
\keyword{counts}
\keyword{values}
