% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_performance.R
\name{pred_acc}
\alias{pred_acc}
\title{Calculate statiscs of accuracy to regression models}
\usage{
pred_acc(obs, pred)
}
\arguments{
\item{obs}{vector with observated data}

\item{pred}{vector with predicted value}
}
\description{
This function Calculate statiscs of accuracy to regression models
}
\details{
Calculate statiscs of accuracy to regression models

details
}
\examples{
\dontrun{
v.obs = c(0,25,75,100)
v.pred = c(1.24,70,97)
pred_acc(v.obs, v.pred)
}
}
