% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_models_performance.R
\name{run_models_performance}
\alias{run_models_performance}
\title{Run Models Performance}
\usage{
run_models_performance(fit.run.model, df.valida, verbose = FALSE)
}
\arguments{
\item{fit.run.model}{list generated by function run_models}

\item{df.valida}{validation data frame}

\item{verbose}{plot graphics of comparation between models}
}
\description{
This function performs the summary of perfomance of models fitter by run_models function
}
\details{
Run Models Performance

details
}
\examples{
\dontrun{
library(dplyr)
vf = labgeo::factor_detect(c_stock, 40)
df = c_stock \%>\%  mutate_at(vf,funs(factor))
dp = data_preparation(df = df, p = 0.10, prune = 0.99)
treino = dp$treino
teste = dp$teste
nm = names(treino)[1]
vrf = recursive_feature_elimination(treino,sizes = c(5,10,15,20,25,30,40),
                                    nfolds = 5,cpu_cores = 7, metric = "Rsquared")
                        cpu_cores = 7, tune_length = 5, metric = "Rsquared")
}
}
