% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{to_table}
\alias{to_table}
\title{transforma a vector of strings in a table do be printed}
\usage{
to_table(txt, n.col, num.digits = 3)
}
\arguments{
\item{txt}{vector of characters (strings)}

\item{n.col}{number  of columns to be created}

\item{num.digits}{number of decimal places to numeric values}
}
