% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc2.R
\name{train_test}
\alias{train_test}
\title{train test split}
\usage{
train_test(df, y, p = 0.75, groups = min(5, length(y)), seed = NULL)
}
\arguments{
\item{df}{dataframe}

\item{y}{outcome variable}

\item{p}{proportion between  train and test sets}

\item{groups}{numbre of groups to be used in stratified sample}

\item{seed}{seed}
}
\description{
Train Test split
}
